backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pypdf2==3.0.1
langchain==0.3.7
langchain-groq==0.2.1
langchain-huggingface==0.1.2
langchain-community==0.3.5
faiss-cpu==1.9.0
python-dotenv==1.0.1
sentence-transformers==3.3.1
